Add Task#do! and Task#undo
I also fixed a very sneaky bug with the calculation of the @done field - Nil Guards and booleans don't fit well together.

Basically, if @done was false, then it got re-calculated at every access.

The problem was invisible, but it becomes visible as soon as you admit changing @done along the way, as this commit does.

I changed the code to something less elegant, but working.