Cluster related specs together
There are a lot of tests in the Task specs. It was getting hard for me to make sense of what was going on.
I clustered related tests together to make it easier to see the functional areas.
There are a few duplicated or unclear tests, but I didn't change that - I just put them together.