Extract class method regexes out of the task class
It makes sense to extract the syntax/tokenizer/splitter to its own
component. This is the first step in a refactoring which will result in
a fair chunk of the parsing and extraction code being nicely separated
from the core list and task models. As a consequence, it will be easier
to fix a couple of weird bugs with the `task.done?` checks.