Adds priority increment / decrement
These functions set and update priorities. They move letters
up and down by one, based on the ordinal alphabet, based
on the todo.txt spec.

If a task does not have a priority, increment sets it to a
default value of (A), while decrement makes no change.
This pattern is used in other apps like todotxt.net,
which has a large number of users. Because of this wide
adoption, these functions follow the same design.

The inc/dec functions return the new value of the item.
This can be used to check if the value has changed without
having to call #priority on the instance again.