Ensure `List#save!` method works correctly
Iterate through the lines using `IO#puts` so that platform-specific line
breaks are supported (this is untested, but should work).

Also adds a check to ensure that a file actually exists before trying to
write to it.

The test is a bit messy, but this can be tidied up properly when the
file handling is extracted to a separate IO module.