Ensure that the root module name matches RubyGems file path convention
Adds a `lib/todo.rb` to match the root module namespace. This ensures
that all code loaded under that module is accessible through the
predictable convention that Gems should follow. Also moves some of the
module level code out of the options file to help make the module-level
organisation and docs more consistent.